moved src/profile.ts to src/settings/
